requirements.txt
# PubMed Search Module Dependencies
# Install with: pip install -r requirements.txt

# Core dependency for PubMed access
biopython>=1.79

# Google AI/Gemini API dependencies
google-generativeai>=0.3.0
google-api-core>=2.11.0
google-genai

# Data handling and analysis (optional but recommended)
pandas>=1.3.0
numpy>=1.21.0

# For enhanced CSV handling (optional)
openpyxl>=3.0.0

# For environment variable management (optional)
python-dotenv>=0.19.0

# DOI downloader dependencies
scidownl
beautifulsoup4>=4.9.0
requests>=2.25.0